Skip scan test if test backend not available

+++ t/01_scan_sane.t
@@ -25,28 +25,40 @@ use Locale::gettext 1.05;    # For translations
 our $d = Locale::gettext->domain($prog_name);
 Gscan2pdf::Frontend::Sane->setup( $prog_name, $d, $logger );
 
-Gscan2pdf::Frontend::Sane->open_device(
- 'test',
- sub { },
- sub { },
- sub {
-  Gscan2pdf::Frontend::Sane->scan_pages(
-   '.',
-   'out%d.pnm',
-   1, 1, 1,
-   sub { },
-   sub { },
-   sub { },
-   sub {
-    is( -s 'out1.pnm', 30807, 'PNM created with expected size' );
-    Gtk2->main_quit;
-   }
-  );
- }
-);
-Gtk2->main;
+SKIP: {
+ if (
+  eval {
+   Gscan2pdf::Frontend::Sane->open_device(
+    'test',
+    sub { },
+    sub { },
+    sub {
+     Gscan2pdf::Frontend::Sane->scan_pages(
+      '.',
+      'out%d.pnm',
+      1, 1, 1,
+      sub { },
+      sub { },
+      sub { },
+      sub {
+       is( -s 'out1.pnm', 30807, 'PNM created with expected size' );
+       Gtk2->main_quit;
+      }
+     );
+    }
+   );
+  }
+   )
+ {
+  Gtk2->main;
 
 #########################
 
-unlink 'out1.pnm';
+  unlink 'out1.pnm';
+ }
+ else {
+  skip 'SANE test backend not installed', 1;
+ }
+}
+
 Gscan2pdf::Frontend::Sane->quit();
